requirement.txt
fastapi
uvicorn
tensorflow-cpu
pillow
python-multipart